+++ RememberThis/Controllers/RememberThisController.cs
@@ -18,15 +18,18 @@ public class RememberThisController : ControllerBase
     private readonly IConfiguration _config;
     private readonly BlobStorage _BlobStorage;
     private readonly SqlDb _SqlDb;
+    private readonly ImageService _ImageService;
 
     private string apiReturnMsg = "Controller Start";
-    private string[] permittedExtensions = new string[] { ".gif", ".png", ".jpg", ".jpeg" };
-    public RememberThisController(ILogger<RememberThisController> logger, IConfiguration config, BlobStorage BlobStorage, SqlDb SqlDb)
+
+    public RememberThisController(ILogger<RememberThisController> logger, IConfiguration config,
+            BlobStorage BlobStorage, SqlDb SqlDb, ImageService ImageService)
     {
         _logger = logger;
         _config = config;
         _BlobStorage = BlobStorage;
         _SqlDb = SqlDb;
+        _ImageService = ImageService;
 
     }
 
@@ -96,7 +99,7 @@ public class RememberThisController : ControllerBase
 
         await stream.CopyToAsync(ms);
 
-        if (IsValidFileExtensionAndSignature(formFile.FileName, ms))
+        if (_ImageService.IsValidFileExtensionAndSignature(formFile.FileName, ms))
         {
             string StorageErrorOrFileName = await _BlobStorage.WritetoAzureStorageAsync(ms, unsafeFileNameAndExt);
 
@@ -142,75 +145,6 @@ public class RememberThisController : ControllerBase
 
 
 
-    public bool IsValidFileExtensionAndSignature(string fileName, MemoryStream streamParam)
-    {
-
-        // might need to go somehwere higher in stack
-        apiReturnMsg = "Validation Method Start";
-        apiReturnMsg = "file check start";
-
-        MemoryStream data = new MemoryStream();
-        streamParam.Position = 0;
-        streamParam.CopyTo(data);
-
-        // this is generally checked by the file upload control itself - but we can double check it here
-        if (data == null || data.Length == 0)
-        {
-            apiReturnMsg = "file empty";
-            return false;
-        }
-
-        var filenameonly = Path.GetFileNameWithoutExtension(fileName);
-        if (string.IsNullOrEmpty(filenameonly))
-        {
-            apiReturnMsg = "file name not valid";
-            return false;
-        }
-
-        var ext = Path.GetExtension(fileName).ToLowerInvariant();
-        if (string.IsNullOrEmpty(ext) || !permittedExtensions.Contains(ext))
-        {
-            apiReturnMsg = "file extension not valid";
-            return false;
-        }
-
-        data.Position = 0;
-
-        using (var reader = new BinaryReader(data))
-        {
-            var signatures = _fileSignature[ext];
-            var headerBytes = reader.ReadBytes(signatures.Max(m => m.Length));
-
-            bool fileSigCorrect = signatures.Any(signature =>
-                headerBytes.Take(signature.Length).SequenceEqual(signature));
-
-            apiReturnMsg = fileSigCorrect ? "file check good" : "file signiture invalid";
-
-            return fileSigCorrect;
-
-        }
-
-    } // End IsValidFileExtensionAndSignature
-
-    private static readonly Dictionary<string, List<byte[]>> _fileSignature = new Dictionary<string, List<byte[]>>
-        {
-            { ".gif", new List<byte[]> { new byte[] { 0x47, 0x49, 0x46, 0x38 } } },
-            { ".png", new List<byte[]> { new byte[] { 0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A } } },
-            { ".jpeg", new List<byte[]>
-                {
-                    new byte[] { 0xFF, 0xD8, 0xFF, 0xE0 },
-                    new byte[] { 0xFF, 0xD8, 0xFF, 0xE2 },
-                    new byte[] { 0xFF, 0xD8, 0xFF, 0xE3 },
-                }
-            },
-            { ".jpg", new List<byte[]>
-                {
-                    new byte[] { 0xFF, 0xD8, 0xFF, 0xE0 },
-                    new byte[] { 0xFF, 0xD8, 0xFF, 0xE1 },
-                    new byte[] { 0xFF, 0xD8, 0xFF, 0xE8 },
-                }
-            }
-        };
 
 
 } // end class rememberthis
