+++ RememberThis/Program.cs
@@ -10,6 +10,7 @@ builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 builder.Services.AddTransient<BlobStorage>();
 builder.Services.AddTransient<SqlDb>();
+builder.Services.AddTransient<ImageService>();
 
 
 var app = builder.Build();
@@ -20,7 +21,7 @@ if (app.Environment.IsDevelopment())
     app.UseSwagger();
     app.UseSwaggerUI();
 }
-else
+else 
 {
     app.UseHttpsRedirection();
 }
